Getting ready for adding GlobRead task
